Refactored code to be able to package game.
