+++ CLI.py
@@ -1,0 +1,4 @@
+from __init__ import main
+
+if __name__ == '__main__':
+    main()
